Add required param gpu_partition_size
As of google provider 3.90+, this parameter is required in the
guest_accelerator block. The default is null so it won't affect existing
configurations.

This is basically the same change that has already landed upstream but
as we cannot upgrade yet, this just backports it for the
beta-public-cluster-update-variant that we are using.